requirements.txt
asttokens==2.0.4
attrs==20.3.0
audioread==2.1.9
colorama==0.4.4
cycler==0.10.0
executing==0.5.4
frosch==0.1.7
iniconfig==1.1.1
install==1.3.4
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
packaging==20.8
Pillow==8.0.1
pluggy==0.13.1
pprintpp==0.4.0
pure-eval==0.1.1
py==1.10.0
pygame==2.0.1
Pygments==2.7.3
pyparsing==2.4.7
pytest==6.2.1
pytest-sugar==0.9.4
python-dateutil==2.8.1
SimpleGUICS2Pygame==2.1.0
SimpleGUITk==1.1.3
six==1.15.0
stack-data==0.1.0
termcolor==1.1.0
toml==0.10.2
yapf==0.30.0
